% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy-solve.R
\name{easy_solve}
\alias{easy_solve}
\title{Solve easy sudokus}
\usage{
easy_solve(board)
}
\arguments{
\item{board}{The sudoku grid in form of a numeric matrix you want to solve.}
}
\value{
if sudoku is solvable it returns TRUE
}
\description{
This function is designed strictly to solve easy sudokus ie. ones with more than
36 clues.
Even though it might be faster than \code{solve_sudoku()}, it has one major drawback:
it doesn't know if a sudoku is solvable or not.
Use carefully.
}
\examples{
easy_solve(given_board)
}
\seealso{
\code{solve_sudoku()}
}
