% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-empty.R
\name{find_empty}
\alias{find_empty}
\title{Find the first empty cell in a sudoku grid}
\usage{
find_empty(board)
}
\arguments{
\item{board}{A sudoku board/grid.}
}
\value{
A numeric vector giving the row and column of the empty cell. Returns NULL
if all cells are filled.
}
\description{
This function finds the next empty/unfilled cell in a sudoku grid. It starts from the
left topmost cell through all cells in that row then onto the next row and so on.
}
\examples{
find_empty(sample_boards()[[1]])

}
