% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_box.R
\name{get_box}
\alias{get_box}
\title{Get the smaller 3 by 3 grid in a sudoku}
\usage{
get_box(board, row, col)
}
\arguments{
\item{board}{A sudoku grid/board}

\item{row}{specified row}

\item{col}{specified column}
}
\value{
A 3 by 3 numeric matrix
}
\description{
This function returns the 3 by 3 grid  on which specified row and column lie on a sudoku
}
\examples{
b1 <- sample_boards()[[2]]
get_box(b1, 4, 6)
}
