% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_board.R
\name{print_board}
\alias{print_board}
\title{Pretty print a sudoku board}
\usage{
print_board(board)
}
\arguments{
\item{board}{A sudoku board/grid.}
}
\description{
This function prints a 9 by 9 sudoku board in a way that is appealing to the eye,
rather than an R matrix.
}
\examples{
# sample board:
board <- matrix(data =
                 c(5, 6, 0, 8, 4, 7, 0, 0, 0,
                   3, 0, 9, 0, 0, 0, 6, 0, 0,
                   0, 0, 8, 0, 0, 0, 0, 0, 0,
                   0, 1, 0, 0, 8, 0, 0, 4, 0,
                   7, 9, 0, 6, 0, 2, 0, 1, 8,
                   0, 5, 0, 0, 3, 0, 0, 9, 0,
                   0, 0, 0, 0, 0, 0, 2, 0, 0,
                   0, 0, 6, 0, 0, 0, 8, 0, 7,
                   0, 0, 0, 3, 1, 6, 0, 5, 9
                 ),
               ncol = 9)
print_board(board)
}
