% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_sudoku}
\alias{solve_sudoku}
\title{Solve a sudoku}
\usage{
solve_sudoku(board)
}
\arguments{
\item{board}{The sudoku grid in form of a numeric matrix you want to solve.}
}
\value{
TRUE or FALSE indicating if the given board is solvable or not
}
\description{
This function solves a given sudoku board. If a sudoku is insolvable it returns
FALSE
}
\examples{
b3 <- sample_boards()[[1]]
print_board(b3)
solve_sudoku(b3)
}
\seealso{
sample_boards()
}
