% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{check if a number can be placed in a certain position.}
\usage{
valid(board, num, row, col)
}
\arguments{
\item{board}{A 9 by 9 sudoku grid}

\item{num}{A number in 1:9}

\item{row}{specified row}

\item{col}{specified column}
}
\value{
TRUE or FALSE indicating if placement is valid or not.
}
\description{
This function checks if the provided number can be placed in the specified row and
column in adherence to the rules of the sudoku game.
}
\examples{
b2 <- sample_boards()[[5]]
valid(b2, 7, 4, 8)
}
