% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_solve.R
\name{easy_solve}
\alias{easy_solve}
\title{Solve easy sudokus}
\usage{
easy_solve(board)
}
\arguments{
\item{board}{The 9 by 9 sudoku grid in form of a numeric matrix.}
}
\value{
if sudoku is solvable it returns TRUE
}
\description{
This function is designed strictly to solve easy sudokus ie. ones with more than
36 clues, but it can sometimes solve medium hard sudokus (27 to 36 clues).
}
\details{
It solves the sudoku deterministically and might be faster than \code{solve_sudoku()}, but
it has two major drawbacks:
can't determine if a sudoku  is solvable and can't also determine if it can solve
a given sudoku. If it takes more than 2 seconds please use \code{solve_sudoku()}.

Use carefully.
}
\examples{
b2 <- sample_boards()[[2]]
easy_solve(b2)
}
\seealso{
\code{solve_sudoku()}
}
