% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-boards.R
\name{sample_boards}
\alias{sample_boards}
\title{Sample sudoku grids}
\usage{
sample_boards()
}
\value{
a list of boards
}
\description{
This function displays a list of 8 sudoku boards which you can try to solve.
}
\examples{
b1 <- sample_boards()[[1]]
b1

sample_boards()[c(1:5)]

}
