% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve_sudoku}
\alias{solve_sudoku}
\title{Solve a sudoku}
\usage{
solve_sudoku(board)
}
\arguments{
\item{board}{The sudoku grid in form of a numeric matrix you want to solve.}
}
\value{
Logical vector of length 1. FALSE if the sudoku board is insolvable.
}
\description{
This function solves any 9 by 9 sudoku board.
}
\examples{
b3 <- sample_boards()[[1]]
print_board(b3)
solve_sudoku(b3)
}
\seealso{
\code{sample_boards()}, \code{easy_solve()}
}
