% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sudoku.R
\name{solve_sudoku}
\alias{solve_sudoku}
\title{Solve a sudoku}
\usage{
solve_sudoku(board)
}
\arguments{
\item{board}{The 9 by 9 sudoku grid in form of a numeric matrix.}
}
\value{
Invisibly returns a logical vector of length 1.
\code{FALSE} if the sudoku board is insolvable.
}
\description{
This function solves any 9 by 9 sudoku board.
}
\examples{
b3 <- sample_boards()[[1]]
print_board(b3)
solve_sudoku(b3)
}
\seealso{
\code{sample_boards()}, \code{easy_solve()}
}
